requirements.txt
agno
python-dotenv
yfinance
groq
duckduckgo-search
fastapi
uvicorn